# swagger/doc/cms.json
{
	"swagger": "2.0",
	"info": {
		"title": "CMS",
		"version": "1",
		"description": "Response format: #definitions/Response"
	},
	"basePath": "/api/v1",
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"security": [
		{
			"JWT": []
		}
	],
...
	"paths": {
...
		"/messages/{messageId}/approve": {
			"post": {
				"tags": [
					"messages"
				],
				"operationId": "approve",
				"summary": "Chấp nhận inquiry",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"description": "Messages ID",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
		"/promotion/ruleset": {
			"get": {
				"tags": [
					"promotions"
				],
				"operationId": "getPromotions",
				"summary": "Get Promotions",
				"produces": [
					"application/json"
				],
				"parameters": [],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PromotionRuleSet"
							}
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			},
			"post": {
				"tags": [
					"promotions"
				],
				"operationId": "createPromotions",
				"summary": "Create Promotions",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PromotionRuleSet"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/PromotionRuleSet"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/promotion/ruleset/{rulesetId}": {
			"get": {
				"tags": [
					"promotions"
				],
				"operationId": "getPromotion",
				"summary": "Get Promotion",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "rulesetId",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/PromotionRuleSet"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			},
			"put": {
				"tags": [
					"promotions"
				],
				"operationId": "updatePromotion",
				"summary": "Update Promotion",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "rulesetId",
						"required": true
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PromotionRuleSet"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/PromotionRuleSet"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			},
			"delete": {
				"tags": [
					"promotions"
				],
				"operationId": "deletePromotion",
				"summary": "Delete Promotion",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "rulesetId",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/promotion/ruleset/{rulesetId}/active": {
			"get": {
				"tags": [
					"promotions"
				],
				"operationId": "activePromotion",
				"summary": "Active / deactive Promotion",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "rulesetId",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/promotion/ruleset_block": {
			"post": {
				"tags": [
					"promotions"
				],
				"operationId": "updatePromotionRulesetBlock",
				"summary": "Update PromotionRulesetBlock",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PromotionRuleSetBlock"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/PromotionRuleSetBlock"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
		"/promotion/ruleset_block/{blockId}": {
			"get": {
				"tags": [
					"promotions"
				],
				"operationId": "getPromotionRulesetBlocks",
				"summary": "Get PromotionRulesetBlocks",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "blockId",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PromotionRuleSetBlock"
							}
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			},
			"delete": {
				"tags": [
					"promotions"
				],
				"operationId": "deletePromotionRulesetBlocks",
				"summary": "Delete PromotionRulesetBlocks",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "blockId",
						"required": true
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"type": "object",
							"properties": {
								"rulesetId": {
									"type": "string"
								}
							}
						},
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
		"/payment/": {
			"post": {
				"tags": [
					"payment"
				],
				"operationId": "createPayment",
				"summary": "Tạo khoản chi",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Payout"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success",
						"schema": {
							"$ref": "#/definitions/Payout"
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
	},
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"in": "header",
			"name": "x-access-token"
		}
	},
	"definitions": {
...
		"Payout": {
			"type": "object",
			"properties": {
				"payoutType": {
					"type": "string"
				},
				"distribute": {
					"type": "boolean"
				},
				"distributeEnd": {
					"type": "date"
				},
				"currencyAmount": {
					"type": "object",
					"properties": {
						"amount": {
							"type": "number"
						},
						"currency": {
							"type": "string"
						},
						"exchange": {
							"type": "number"
						},
						"feePercent": {
							"type": "number"
						},
						"exchangedAmount": {
							"type": "number"
						}
					}
				},
				"description": {
					"type": "string"
				},
				"state": {
					"type": "string",
					"enum": [
						"processing",
						"transferred",
						"confirmed"
					]
				},
				"source": {
					"type": "string",
					"enum": [
						"banking",
						"cash",
						"swipe_card"
					]
				},
				"collector": {
					"type": "string"
				},
				"collectorCustomName": {
					"type": "string"
				},
				"products": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"currencyAmount": {
								"type": "object",
								"properties": {
									"amount": {
										"type": "number"
									},
									"currency": {
										"type": "string"
									}
								}
							},
							"bookingId": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
		"PromotionRuleSet": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"active": {
					"type": "boolean"
				},
				"airbnb": {
					"type": "object"
				},
				"startDate": {
					"type": "string"
				},
				"endDate": {
					"type": "string"
				},
				"dayOfWeeks": {
					"type": "string"
				},
				"bookStartDate": {
					"type": "string"
				},
				"bookEndDate": {
					"type": "string"
				},
				"bookDayOfWeeks": {
					"type": "string"
				},
				"activeOTAs": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"ratio": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"otaName": {
								"type": "string"
							},
							"value": {
								"type": "number"
							}
						}
					}
				},
				"tags": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
		"PromotionRuleSetBlock": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"active": {
					"type": "boolean"
				},
				"blockId": {
					"type": "string"
				},
				"rulesetId": {
					"type": "string"
				},
				"promotionIds": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"rateIds": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"tags": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
...
	}
}